table tests except dependency

+++ table.js
@@ -1,6 +1,12 @@
 var DBEnvError = require("./errors");
 
 function Table (name, dao) {
+    if (arguments.length != 2) {
+        throw new DBEnvError("Expect exactly 2 parameters");
+    }
+    if (!dao || typeof dao.insert != "function" || typeof dao.delete != "function") {
+        throw new DBEnvError("Expect valid DAO with insert and delete methods");
+    }
     this.name = name;
     this.fields = {};
     this.dao = dao;
@@ -15,23 +21,21 @@ Table.prototype.addField = function (name, generator, key) {
     if (this.finalized) {
         throw new DBEnvError("Can`t add field to finalized table");
     }
-    var self = this;
-    self.fields[name] = {
-        generator: generator,
-        dependency: generator.dependsOn
-    };
+    this.fields[name] = generator;
     if (key) {
         this.setKey(name);
     }
+    return this;
 };
 
 Table.prototype.insert = function (data) {
     if (!this.finalized) {
         throw new DBEnvError("Can`t insert row into not finalized table");
     }
+    var _data = data || {};
     var self = this;
     var insertObject = Object.keys(this.fields).reduce(function (obj, fieldName) {
-        obj[fieldName] = data.hasOwnProperty(fieldName) ? self.fields[fieldName].constructor(data[fieldName]) : self.fields[fieldName].constructor();
+        obj[fieldName] = _data.hasOwnProperty(fieldName) ? self.fields[fieldName](_data[fieldName]) : self.fields[fieldName]();
         return obj;
     }, {});
     self.dao.insert(insertObject);
@@ -44,6 +48,9 @@ Table.prototype.cleanup = function () {
     if (!this.finalized) {
         throw new DBEnvError("Can`t cleanup not finalized table");
     }
+    if (!Object.keys(this.rows).length) {
+        return;
+    }
     this.dao.delete();
     this.rows = {};
 };
@@ -52,6 +59,9 @@ Table.prototype.setKey = function (fieldName) {
     if (this.finalized) {
         throw new DBEnvError("Can`t set key to finalized table");
     }
+    if (!this.fields[fieldName]) {
+        throw new DBEnvError("Can`t set non existent field as key field");
+    }
     this.key = fieldName;
 };
 
@@ -59,14 +69,17 @@ Table.prototype.getRow = function (key, fields, populated) {
     if (!this.finalized) {
         throw new DBEnvError("Can`t get row from not finalized table");
     }
-    var _fields = fields? fields : Object.keys(this.fields);
+    var _fields = fields ? fields : Object.keys(this.fields);
     var self = this;
     var row = self.rows[key];
+    if (!row) {
+        return undefined;
+    }
     return _fields.reduce(function (obj, field) {
         var fieldName = typeof field == "string" ? field : field.name;
         var dependency = self.fields[fieldName].dependency;
         if (dependency && populated) {
-            obj[fieldName] = dependency.getRow(row[fieldName], field.fields, field.hadOwnProperty(populated) ? field.populated : true);
+            obj[fieldName] = dependency.getRow(row[fieldName], field.fields, field.hadOwnProperty("populated") ? field.populated : true);
         } else {
             obj[fieldName] = row[fieldName];
         }
@@ -83,9 +96,10 @@ Table.prototype.finalize = function () {
         throw new DBEnvError("Can`t finalize finalized table");
     }
     if (Object.keys(this.fields).length === 0) {
-        throw new DBEnvError("Can`t finalize table without rows");
+        throw new DBEnvError("Can`t finalize table without fields");
     }
-    if (!this.key || !this.fields[key]) {
+    var key = this.key;
+    if (!key || !this.fields[key]) {
         throw new DBEnvError("Can`t finalize table without key field");
     }
     return this.finalized = true;
